+++ packages/layout/src/node-renderers/split-node-renderer/SplitLayoutPanelContent.tsx
@@ -2,7 +2,6 @@ import {FC} from 'react';
 import {useSplitNodeContext} from '../../LayoutNodeContext';
 import {LayoutNode} from '@sqlrooms/layout-config';
 import {LayoutNodeRenderer} from '../LayoutNodeRenderer';
-import {appendSemanticLayoutPath} from '../utils';
 
 type SplitLayoutPanelContentProps = {
   node: LayoutNode;
@@ -13,12 +12,10 @@ export const SplitLayoutPanelContent: FC<SplitLayoutPanelContentProps> = ({
 }) => {
   const {node: parentNode, path} = useSplitNodeContext();
 
-  const nextPath = appendSemanticLayoutPath(path, node);
-
   return (
     <LayoutNodeRenderer
       node={node}
-      path={nextPath}
+      path={path}
       containerType="split"
       containerId={parentNode.id}
       parentDirection={parentNode.direction}